attempt to sketch engine and lots of subsequent random stuff. sorry Dan
