+++ validator/src/validator/logging_utils.py
@@ -1,25 +1,26 @@
 import logging
 from logging.handlers import RotatingFileHandler
-from dotenv import load_dotenv
 import os
 
 
 def setup_logging(logger):
-	load_dotenv()
 
 	log_file = os.getenv("LOG_FILE", "logs/validate.log")
 	error_file = os.getenv("ERROR_FILE", "logs/validate.err")
-	log_level = os.getenv("LOG_LEVEL", "INFO").upper()
-
-	logger.setLevel(getattr(logging, log_level, logging.INFO))
+	
+	log_level = os.getenv("LOG_LEVEL", "DEBUG").upper()
+	logger.setLevel(getattr(logging, log_level, logging.DEBUG))
 
 	file_handler = RotatingFileHandler(log_file, maxBytes=5 * 1024 * 1024, backupCount=3)
+	file_handler.setLevel(logging.DEBUG)
+
 	error_handler = logging.FileHandler(error_file, "w", encoding="utf-8")
 	error_handler.setLevel(logging.ERROR)
 
 	console_handler = logging.StreamHandler()
 	console_handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
-	console_handler.setLevel(logging.INFO)
+	console_handler.setLevel(logging.DEBUG)
+
 	logger.addHandler(console_handler)
 
 	formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
@@ -32,8 +33,8 @@ def setup_logging(logger):
 
 def pprint(args):
 
-    ret_str = ""
-    for key, value in args.items():
-        ret_str += f"{key:40} - {str(value):80}\n"
+	ret_str = ""
+	for key, value in args.items():
+		ret_str += f"{key:40} - {str(value):80}\n"
 
-    return ret_str
+	return ret_str
