--- validator/src/validator/validator.py
@@ -1,1038 +1,0 @@
-import collections
-import regex as re
-
-from validator.incident import Error, Warning, TestClass
-import validator.utils as utils
-import validator.compiled_regex as crex
-
-def validate_mwt_empty_vals(cols):
-	"""
-	Checks that a multi-word token has _ empty values in all fields except MISC.
-	This is required by UD guidelines although it is not a problem in general,
-	therefore a level 2 test.
-
-	Parameters
-	----------
-	cols : list
-		The values of the columns on the current node / token line.
-	"""
-
-	if not utils.is_multiword_token(cols):
-		ret = Error(level=0,
-					testclass=TestClass.INTERNAL,
-					testid='internal error')
-		return [ret]
-
-	# all columns except the first two (ID, FORM) and the last one (MISC)
-	for col_idx in range(utils.LEMMA, utils.MISC):
-
-		# Exception: The feature Typo=Yes may occur in FEATS of a multi-word token.
-		if cols[col_idx] != '_' and (col_idx != utils.FEATS or cols[col_idx] not in ['Typo=Yes', '_']):
-			ret = Error(level=2,
-						testclass=TestClass.FORMAT,
-						testid='mwt-nonempty-field',
-						message=f"A multi-word token line must have '_' in the column {utils.COLNAMES[col_idx]}. Now: '{cols[col_idx]}'."
-						)
-			return [ret]
-
-	return []
-
-
-def validate_empty_node_empty_vals(cols):
-	"""
-	Checks that an empty node has _ empty values in HEAD and DEPREL. This is
-	required by UD guidelines but not necessarily by CoNLL-U, therefore
-	a level 2 test.
-
-	Parameters
-	----------
-	cols : list
-		The values of the columns on the current node / token line.
-	"""
-	if not utils.is_empty_node(cols):
-		ret = Error(level=0,
-					testclass=TestClass.INTERNAL,
-					testid='internal error')
-		return [ret]
-
-	ret = []
-	for col_idx in (utils.HEAD, utils.DEPREL): # ! isn't it worth it to check also DEPS here?
-		if cols[col_idx]!= '_':
-			ret.append(
-					Error(
-					level=2,
-					testclass=TestClass.FORMAT,
-					testid='mwt-nonempty-field',
-					message=f"An empty node must have '_' in the column {utils.COLNAMES[col_idx]}. Now: '{cols[col_idx]}'."
-					)
-			)
-	return ret
-
-#! proposal: rename into validate_deps_deprel_contraints, or also check UPOS format (not value)
-#! I don't like that it relies on crex
-def validate_character_constraints(cols):
-	"""
-	Checks general constraints on valid characters, e.g. that UPOS
-	only contains [A-Z].
-
-	Parameters
-	----------
-	cols : list
-		The values of the columns on the current node / token line.
-	"""
-
-	if utils.is_multiword_token(cols):
-		return []
-
-	# Do not test the regular expression crex.upos here. We will test UPOS
-	# directly against the list of known tags. That is a level 2 test, too.
-
-	if utils.is_empty_node(cols) and cols[utils.DEPREL] == '_':
-		return []
-
-	ret = []
-	if not crex.deprel.fullmatch(cols[utils.DEPREL]):
-		ret.append(
-			Error(
-				level=2,
-				testclass=TestClass.SYNTAX,
-				testid='invalid-deprel',
-				message=f"Invalid DEPREL value '{cols[utils.DEPREL]}'. Only lowercase"
-			)
-		)
-
-	try:
-		deps = utils.deps_list(cols)
-	except ValueError:
-		ret.append(
-			Error(
-					level=2,
-					testclass=TestClass.ENHANCED,
-					testid='invalid-deps',
-					message=f"Failed to parse DEPS: '{cols[utils.DEPS]}'."
-			)
-		)
-		return ret
-
-	for _, edep in deps:
-		if not crex.edeprel.fullmatch(edep):
-			ret.append(
-				Error(
-					level=2,
-					testclass=TestClass.ENHANCED,
-					testid='invalid-edeprel',
-					message=f"Invalid enhanced relation type: '{edep}' in '{cols[utils.DEPS]}'."
-				)
-			)
-	return ret
-
-def validate_upos(cols, Fspecs):
-	"""
-	Checks that the UPOS field contains one of the 17 known tags.
-
-	Parameters
-	----------
-	cols : list
-		The values of the columns on the current node / token line.
-	Fspecs : UDSpecs
-		The object containing specific information about the allowed values
-	"""
-
-	#! added checking for mwt?
-	if utils.is_multiword_token(cols) and cols[utils.UPOS] == '_':
-		return []
-
-	if utils.is_empty_node(cols) and cols[utils.UPOS] == '_':
-		return []
-
-	# Just in case, we still match UPOS against the regular expression that
-	# checks general character constraints. However, the list of UPOS, loaded
-	# from a JSON file, should conform to the regular expression.
-	ret = []
-	print(cols[utils.UPOS], crex.upos.fullmatch(cols[utils.UPOS]))
-	if not crex.upos.fullmatch(cols[utils.UPOS]) or cols[utils.UPOS] not in Fspecs.upos:
-		ret.append(
-			Error(
-	  			level=2,
-				testclass=TestClass.MORPHO,
-				testid='unknown-upos',
-				message=f"Unknown UPOS tag: '{cols[utils.UPOS]}'."
-			)
-		)
-
-	return ret
-
-# ! proposal: rename into feature format or something alike
-def validate_features_level2(cols):
-	"""
-	Checks general constraints on feature-value format: Permitted characters in
-	feature name and value, features must be sorted alphabetically, features
-	cannot be repeated etc.
-
-	Parameters
-	----------
-	cols : list
-		The values of the columns on the current node / token line.
-
-	"""
-	ret = []
-
-	feats = cols[utils.FEATS]
-	if feats == '_':
-		return ret
-
-	# self.features_present(state) # TODO: do elsewhere
-
-	feat_list = feats.split('|') #! why not a function in utils? Like the one that gets deps
-	if [f.lower() for f in feat_list] != list(sorted(f.lower() for f in feat_list)):
-		ret.append(
-			Error(
-				level=2,
-				testclass=TestClass.MORPHO,
-				testid='unsorted-features',
-				message=f"Morphological features must be alphabetically sorted: '{feats}'."
-			)
-		)
-
-	# I'll gather the set of features here to check later that none is repeated.
-	attr_set = set()
-	# Level 2 tests character properties and canonical order but not that the f-v pair is known.
-
-	for feat_val in feat_list:
-		match = crex.featval.fullmatch(feat_val)
-		if not match:
-			ret.append(
-				Error(
-					level=2,
-					testclass=TestClass.MORPHO,
-					testid='invalid-feature',
-					message=f"Spurious morphological feature: '{feat_val}'. Should be of the form Feature=Value and must start with [A-Z] and only contain [A-Za-z0-9]."
-				)
-			)
-
-			# to prevent misleading error "Repeated features are disallowed"
-			attr_set.add(feat_val)
-
-		else:
-			# Check that the values are sorted as well
-			attr = match.group(1)
-			attr_set.add(attr)
-			values = match.group(2).split(',')
-			if len(values) != len(set(values)):
-				ret.append(
-					Error(
-						level=2,
-						testclass=TestClass.MORPHO,
-						testid='repeated-feature-value',
-						message=f"Repeated feature values are disallowed: '{feats}' (error generated by feature '{attr}')"
-					)
-				)
-			if [v.lower() for v in values] != sorted(v.lower() for v in values):
-				ret.append(
-					Error(
-						level=2,
-						testclass=TestClass.MORPHO,
-						testid='unsorted-feature-values',
-						message=f"If a feature has multiple values, these must be sorted: '{feat_val}'"
-					)
-				)
-			for v in values:
-				if not crex.val.fullmatch(v): # ! can this ever be true? If val.fullmatch() does not match, than also featval.fullmatch() wouldn't
-					ret.append(
-						Error(
-							level=2,
-							testclass=TestClass.MORPHO,
-							testid='invalid-feature-value',
-							message=f"Spurious value '{v}' in '{feat_val}'. Must start with [A-Z0-9] and only contain [A-Za-z0-9]."
-						)
-					)
-
-	if len(attr_set) != len(feat_list):
-		ret.append(
-			Error(
-				level=2,
-				testclass=TestClass.MORPHO,
-				testid='repeated-feature',
-				message=f"Repeated features are disallowed: '{feats}'."
-			)
-		)
-
-	# Subsequent higher-level tests could fail if a feature is not in the
-	# Feature=Value format. If that happens, we will return False and the caller
-	# can skip the more fragile tests.
-	# TODO: the engine has to know that 'invalid-feature' is a testid that prevents from further
-	return ret
-
-# TODO: write tests
-def validate_deps(cols):
-	"""
-	Validates that DEPS is correctly formatted and that there are no
-	self-loops in DEPS (longer cycles are allowed in enhanced graphs but
-	self-loops are not).
-
-	This function must be run on raw DEPS before it is fed into Udapi because
-	it checks the order of relations, which is not guaranteed to be preserved
-	in Udapi. On the other hand, we assume that it is run after
-	validate_id_references() and only if DEPS is parsable and the head indices
-	in it are OK.
-
-	Parameters
-	----------
-	cols : list
-		The values of the columns on the current node / token line.
-	"""
-
-	# TODO: the engine must assume that it is run after validate_id_references() and only if DEPS is parsable and the head indices in it are OK.
-
-	ret = []
-	if not (utils.is_word(cols) or utils.is_empty_node(cols)):
-		return ret
-
-	# TODO: move elsewhere
-	# Remember whether there is at least one difference between the basic
-	# tree and the enhanced graph in the entire dataset.
-	#if cols[utils.DEPS] != '_' and cols[utils.DEPS] != cols[utils.HEAD]+':'+cols[utils.DEPREL]:
-	#	state.seen_enhancement = line
-
-	# We already know that the contents of DEPS is parsable (deps_list() was
-	# first called from validate_id_references() and the head indices are OK).
-	deps = utils.deps_list(cols)
-	###!!! Float will not work if there are 10 empty nodes between the same two
-	###!!! regular nodes. '1.10' is not equivalent to '1.1'.
-	# ORIGINAL VERSION: heads = [float(h) for h, d in deps]
-
-	# NEW VERSION:
-	#! maybe do this only is [0-9]+.[1-9][0-9]+ is present somewhere?
-	heads = [h for h, _ in deps]
-	floating_len = []
-	for h in heads:
-		if "." in h:
-			floating_len.append(len(h))
-		else:
-			floating_len.append(0)
-	hacked_heads = [h+'00000001' for i, h in enumerate(heads) if floating_len[i]]
-	hacked_heads_sorted = sorted(zip(hacked_heads, floating_len), key= lambda x: float(x[0]))
-	hacked_heads_restored = []
-	for x, y in hacked_heads_sorted:
-		if y:
-			hacked_heads_sorted.append(x[:y])
-		else:
-			hacked_heads_sorted.append(x)
-
-	# if heads != sorted(heads): # sort strings keeping the integer-like ordering
-	if heads != hacked_heads_restored:
-		ret.append(
-			Error(
-				level=2,
-				testclass=TestClass.FORMAT,
-				testid='unsorted-deps',
-				message=f"DEPS not sorted by head index: '{cols[utils.DEPS]}'"
-			)
-		)
-	else:
-		lasth = None
-		lastd = None
-		for h, d in deps:
-			if h == lasth:
-				if d < lastd:
-					ret.append(
-						Error(
-							level=2,
-							testclass=TestClass.FORMAT,
-							testid='unsorted-deps-2',
-							message=f"DEPS pointing to head '{h}' not sorted by relation type: '{cols[utils.DEPS]}'"
-						)
-					)
-				elif d == lastd:
-					ret.append(
-						Error(
-							level=2,
-							testclass=TestClass.FORMAT,
-							testid='unsorted-deps',
-							message=f"DEPS contain multiple instances of the same relation '{h}:{d}'"
-						)
-					)
-			lasth = h
-			lastd = d
-
-	try:
-		id_ = float(cols[utils.ID])
-	except ValueError:
-		# This error has been reported previously.
-		# TODO: check, before there was just a return
-		return ret
-
-	if id_ in heads:
-		ret.append(
-			Error(
-				level=2,
-				testclass=TestClass.ENHANCED,
-				testid='deps-self-loop',
-				message=f"Self-loop in DEPS for '{cols[utils.ID]}'"
-			)
-		)
-
-	return ret
-
-
-def validate_misc(cols):
-	"""
-	In general, the MISC column can contain almost anything. However, if there
-	is a vertical bar character, it is interpreted as the separator of two
-	MISC attributes, which may or may not have the form of attribute=value pair.
-	In general it is not forbidden that the same attribute appears several times
-	with different values, but this should not happen for selected attributes
-	that are described in the UD documentation.
-
-	This function must be run on raw MISC before it is fed into Udapi because
-	Udapi is not prepared for some of the less recommended usages of MISC.
-
-	Parameters
-	----------
-	cols : list
-		The values of the columns on the current node / token line.
-	"""
-	# Incident.default_lineno = line
-	# Incident.default_level = 2
-	# Incident.default_testclass = 'Warning'
-
-	ret = []
-
-	if cols[utils.MISC] == '_':
-		return ret
-
-	misc = [ma.split('=', 1) for ma in cols[utils.MISC].split('|')] #! why not using a function in utils? Just like the one for features
-	mamap = collections.defaultdict(int)
-	for ma in misc:
-		if ma[0] == '':
-			if len(ma) == 1:
-				ret.append(
-					Warning(
-						level=2,
-						testid='empty-misc',
-						message="Empty attribute in MISC; possible misinterpreted vertical bar?"
-					)
-				)
-			else:
-				ret.append(
-					Warning(
-						level=2,
-						testid='empty-misc-key',
-						message=f"Empty MISC attribute name in '{ma[0]}={ma[1]}'."
-					)
-				)
-		# We do not warn about MISC items that do not contain '='.
-		# But the remaining error messages below assume that ma[1] exists.
-		if len(ma) == 1:
-			ma.append('')
-		if re.match(r"^\s", ma[0]) or \
-			re.match(r"\s$", ma[0]) or \
-			re.match(r"^\s", ma[1]) or \
-			re.search(r"\s$", ma[1]):
-			ret.append(
-				level=2,
-				testid='misc-extra-space',
-    			message=f"MISC attribute: leading or trailing extra space in '{'='.join(ma)}'."
-			)
-
-		if re.match(r"^(SpaceAfter|Lang|Translit|LTranslit|Gloss|LId|LDeriv)$", ma[0]):
-			mamap[ma[0]] += 1
-		elif re.match(r"^\s*(spaceafter|lang|translit|ltranslit|gloss|lid|lderiv)\s*$", ma[0], re.IGNORECASE):
-			ret.append(
-				Warning(
-					level=2,
-					testid='misc-attr-typo',
-					message=f"Possible typo (case or spaces) in MISC attribute '{'='.join(ma)}'."
-				)
-
-			)
-
-	for ma in mamap:
-		if mamap[ma] > 1:
-			ret.append(
-				Error(
-					level=2,
-					testclass=TestClass.FORMAT,
-					testid='repeated-misc',
-					message=f"MISC attribute '{ma}' not supposed to occur twice"
-				)
-			)
-
-	return ret
-
-# TODO: write tests
-def validate_deps_all_or_none(sentence, seen_enhanced_graph):
-	"""
-	Takes the list of non-comment lines (line = list of columns) describing
-	a sentence. Checks that enhanced dependencies are present if they were
-	present at another sentence, and absent if they were absent at another
-	sentence.
-	"""
-	ret = []
-	egraph_exists = False # enhanced deps are optional
-	for cols in sentence:
-		# if utils.is_multiword_token(cols):
-		# 	continue
-		if not utils.is_multiword_token(cols) and (utils.is_empty_node(cols) or cols[utils.DEPS] != '_'):
-			egraph_exists = True
-
-	# We are currently testing the existence of enhanced graphs separately for each sentence.
-	# However, we should not allow that one sentence has a connected egraph and another
-	# has no enhanced dependencies. Such inconsistency could come as a nasty surprise
-	# to the users.
-	# Incident.default_lineno = state.sentence_line
-	# Incident.default_level = 2
-	# Incident.default_testclass = 'Enhanced'
-	if egraph_exists:
-		if not seen_enhanced_graph:
-			# TODO: do elsewhere
-			# state.seen_enhanced_graph = state.sentence_line
-			ret.append(
-				Error(
-					testclass=TestClass.ENHANCED,
-					testid='edeps-only-sometimes',
-					message=f"Enhanced graph must be empty because we saw empty DEPS earlier."
-				)
-			)
-			#! we should add something to this message in the engine where we have access to the state:
-			#on line {state.seen_tree_without_enhanced_graph}
-
-	else:
-#		if not state.seen_tree_without_enhanced_graph:
-			# TODO: do elsewhere
-#  			state.seen_tree_without_enhanced_graph = state.sentence_line
-			if seen_enhanced_graph:
-				ret.append(
-					Error(
-						level=2,
-						testid='edeps-only-sometimes',
-						message=f"Enhanced graph cannot be empty because we saw non-empty DEPS earlier."
-					)
-				)
-				#! we should add something to this message in the engine where we have access to the state:
-				#  on line {state.seen_enhanced_graph}
-
-def validate_newlines(inp):
-	"""
-	Checks that the input file consistently uses linux-style newlines (LF only,
-	not CR LF like in Windows). To be run on the input file handle after the
-	whole input has been read.
-	"""
-
-	if inp.newlines and inp.newlines != '\n':
-		ret = Error(
-				level=1,
-				testclass=TestClass.FORMAT,
-				testid='non-unix-newline',
-				message='Only the unix-style LF line terminator is allowed.'
-			)
-		return [ret]
-
-	return []
-
-
-# TODO: move elsewhere
-# # If a multi-word token has Typo=Yes, its component words must not have it.
-# 			# We must remember the span of the MWT and check it in validate_features_level4().
-# 			m = crex.mwtid.fullmatch(cols[ID])
-# 			state.mwt_typo_span_end = m.group(2)
-
-def validate_token_ranges(sentence):
-    """
-    Checks that the word ranges for multiword tokens are valid.
-
-    Parameters
-    ----------
-    sentence : list
-        A list of lists representing a sentence in tabular format.
-
-    Returns
-    -------
-    incidents : list
-        A list of Incidents (empty if validation is successful). 
-    """
-    incidents = []
-    covered = set()
-    for cols in sentence:
-        if not "-" in cols[utils.ID]:
-            continue
-        m = crex.mwtid.fullmatch(cols[utils.ID])
-        if not m: 
-            incidents.append(Error(
-                testid="invalid-word-interval",
-                message=f"Spurious word interval definition: '{cols[utils.ID]}'."
-            ))
-            continue
-        start, end = m.groups()
-        start, end = int(start), int(end)
-        # Do not test if start >= end: 
-        # This is tested in validate_id_sequence().
-        if covered & set(range(start, end+1)):
-            incidents.append(Error(
-                testid='overlapping-word-intervals',
-                message=f'Range overlaps with others: {cols[utils.ID]}'))
-        covered |= set(range(start, end+1))
-    return incidents
-
-def validate_id_sequence(sentence):
-    """
-    Validates that the ID sequence is correctly formed.
-    If this function returns an nonempty list, subsequent tests should not be run.
-
-    Parameters
-    ----------
-    sentence : list
-        A list of lists representing a sentence in tabular format.
-
-    Returns
-    -------
-    incidents : list
-        A list of Incidents (empty if validation is successful). 
-    """
-    incidents = []
-    words=[]
-    tokens=[]
-    current_word_id, next_empty_id = 0, 1
-    for cols in sentence:
-        # Check for the format of the ID value. (ID must not be empty.)
-        if not (utils.is_word(cols) or utils.is_empty_node(cols) or utils.is_multiword_token(cols)):
-            incidents.append(Error(
-                testid='invalid-word-id',
-                message=f"Unexpected ID format '{cols[utils.ID]}'."
-            ))
-            continue
-        if not utils.is_empty_node(cols):
-            next_empty_id = 1    # reset sequence
-        if utils.is_word(cols):
-            t_id = int(cols[utils.ID])
-            current_word_id = t_id
-            words.append(t_id)
-            # Not covered by the previous interval?
-            if not (tokens and tokens[-1][0] <= t_id and tokens[-1][1] >= t_id):
-                tokens.append((t_id, t_id)) # nope - let's make a default interval for it
-
-        # ! looks like a duplicate of validate_id_sequence
-        elif utils.is_multiword_token(cols):
-            match = crex.mwtid.fullmatch(cols[utils.ID]) # Check the interval against the regex
-            if not match: # This should not happen. The function utils.is_multiword_token() would then not return True.
-                incidents.append(Error(
-                    testid='invalid-word-interval',
-                    message=f"Spurious word interval definition: '{cols[utils.ID]}'."
-                ))
-                continue
-            beg, end = int(match.group(1)), int(match.group(2))
-            if not ((not words and beg >= 1) or (words and beg >= words[-1] + 1)):
-                incidents.append(Error(
-                    testid='misplaced-word-interval',
-                    message='Multiword range not before its first word.'
-                ))
-                continue
-            tokens.append((beg, end))
-        elif utils.is_empty_node(cols):
-            word_id, empty_id = (int(i) for i in utils.parse_empty_node_id(cols))
-            if word_id != current_word_id or empty_id != next_empty_id:
-                incidents.append(Error(
-                    testid='misplaced-empty-node',
-                    message=f'Empty node id {cols[utils.ID]}, expected {current_word_id}.{next_empty_id}'
-                ))
-            next_empty_id += 1
-            # Interaction of multiword tokens and empty nodes if there is an empty
-            # node between the first word of a multiword token and the previous word:
-            # This sequence is correct: 4 4.1 5-6 5 6
-            # This sequence is wrong:   4 5-6 4.1 5 6
-            if word_id == current_word_id and tokens and word_id < tokens[-1][0]:
-                incidents.append(Error(
-                    testid='misplaced-empty-node',
-                    message=f"Empty node id {cols[utils.ID]} must occur before multiword token {tokens[-1][0]}-{tokens[-1][1]}."
-                ))
-    # Now let's do some basic sanity checks on the sequences.
-    # Expected sequence of word IDs is 1, 2, ...
-    expstrseq = ','.join(str(x) for x in range(1, len(words) + 1))
-    wrdstrseq = ','.join(str(x) for x in words)
-    if wrdstrseq != expstrseq:
-        incidents.append(Error(
-            testid='word-id-sequence',
-            message=f"Words do not form a sequence. Got '{wrdstrseq}'. Expected '{expstrseq}'."
-        ))
-    # Check elementary sanity of word intervals.
-    # Remember that these are not just multi-word tokens. Here we have intervals even for single-word tokens (b=e)!
-    for (b, e) in tokens:
-        if e < b: # end before beginning
-            incidents.append(Error(
-                testid='reversed-word-interval',
-                message=f'Spurious token interval {b}-{e}'
-            ))
-            continue
-        if b < 1 or e > len(words): # out of range
-            incidents.append(Error(
-                testid='word-interval-out',
-                message=f'Spurious token interval {b}-{e} (out of range)'
-            ))
-            continue
-    return incidents
-
-def validate_id_references(sentence):
-    """
-    Verifies that HEAD and DEPS reference existing IDs. If this function returns a nonempty list, most of the other tests should be skipped for the current
-    sentence (in particular anything that considers the tree structure).
-
-    Parameters
-    ----------
-    sentence : list
-        A list of lists representing a sentence in tabular format.
-
-    Returns
-    -------
-    incidents : list
-        A list of Incidents (empty if validation is successful). 
-    """
-    incidents = []
-    word_tree = [cols for cols in sentence if utils.is_word(cols) or utils.is_empty_node(cols)]
-    ids = set([cols[utils.ID] for cols in word_tree])
-    for cols in word_tree:
-        # Test the basic HEAD only for non-empty nodes.
-        # We have checked elsewhere that it is empty for empty nodes.
-        if not utils.is_empty_node(cols):
-            match = crex.head.fullmatch(cols[utils.HEAD])
-            if match is None:
-                incidents.append(Error(
-                    testid='invalid-head',
-                    message=f"Invalid HEAD: '{cols[utils.HEAD]}'."
-                ))
-            if not (cols[utils.HEAD] in ids or cols[utils.HEAD] == '0'):
-                incidents.append(Error(
-                    testclass=TestClass.SYNTAX,
-                    testid='unknown-head',
-                    message=f"Undefined HEAD (no such ID): '{cols[id.HEAD]}'."
-                ))
-        try:
-            deps = utils.deps_list(cols)
-        except ValueError:
-            # Similar errors have probably been reported earlier.
-            incidents.append(Error(
-                testid='invalid-deps',
-                message=f"Failed to parse DEPS: '{cols[utils.DEPS]}'."
-            ))
-            continue
-        for head, _ in deps:
-            match = crex.ehead.fullmatch(head)
-            if match is None:
-                incidents.append(Error(
-                    testid='invalid-ehead',
-                    message=f"Invalid enhanced head reference: '{head}'."
-                ))
-            if not (head in ids or head == '0'):
-                incidents.append(Error(
-                    testclass=TestClass.ENHANCED,
-                    testid='unknown-ehead',
-                    message=f"Undefined enhanced head reference (no such ID): '{head}'."
-                ))
-    return incidents
-
-def validate_tree(sentence, node_line, single_root):
-    """
-    Performs basic validation of the tree structure.
-
-    This function originally served to build a data structure that would
-    describe the tree and make it accessible during subsequent tests. Now we
-    use the Udapi data structures instead but we still have to call this
-    function first because it will survive and report ill-formed input. In
-    such a case, the Udapi data structure will not be built and Udapi-based
-    tests will be skipped.
-
-    This function should be called only if both ID and HEAD values have been
-    found valid for all tree nodes, including the sequence of IDs and the references from HEAD to existing IDs.
-
-    Parameters
-    ----------
-    sentence : list
-        A list of lists representing a sentence in tabular format.
-    node_line : int
-        A file-wide line counter.
-    single_root : bool
-        A flag indicating whether we should check that there is a single root.
-
-    Returns
-    -------
-    incidents : list
-        A list of Incidents (empty if validation is successful). 
-    """
-    # node_line = state.sentence_line - 1 TODO: this should be done by the engine
-    incidents = []
-    children = {} # int(node id) -> set of children
-    n_words = 0
-    for cols in sentence:
-        node_line += 1
-        if not utils.is_word(cols):
-            continue
-        n_words += 1
-        # ID and HEAD values have been validated before and this function would
-        # not be called if they were not OK. So we can now safely convert them
-        # to integers.
-        id_ = int(cols[utils.ID])
-        head = int(cols[utils.HEAD])
-        if head == id_:
-            incidents.append(Error(
-                testclass=TestClass.SYNTAX,
-                lineno=node_line,
-                testid='head-self-loop',
-                message=f'HEAD == ID for {cols[utils.ID]}'
-            ))
-        # Incrementally build the set of children of every node.
-        children.setdefault(head, set()).add(id_)
-    word_ids = list(range(1, n_words+1))
-    # Check that there is just one node with the root relation.
-    children_0 = sorted(children.get(0, []))
-    if len(children_0) > 1 and single_root:
-        incidents.append(Error(
-            testclass=TestClass.SYNTAX,
-            testid='multiple-roots',
-            message=f"Multiple root words: {children_0}"
-        ))
-    projection = set()
-    node_id = 0
-    nodes = list((node_id,))
-    while nodes:
-        node_id = nodes.pop()
-        children_id = sorted(children.get(node_id, []))
-        for child in children_id:
-            if child in projection:
-                continue # skip cycles
-            projection.add(child)
-            nodes.append(child)
-    unreachable = set(word_ids) - projection
-    if unreachable:
-        str_unreachable = ','.join(str(w) for w in sorted(unreachable))
-        incidents.append(Error(
-            testclass=TestClass.SYNTAX,
-            testid='non-tree',
-            message=f'Non-tree structure. Words {str_unreachable} are not reachable from the root 0.'
-        ))
-    return incidents
-
-def validate_sent_id(comments, allow_slash, known_sent_ids):
-    """
-    Checks that sentence id exists, is well-formed and unique.
-    
-    Parameters
-    ----------
-    comments : list
-        A list of comments, represented as strings.
-    allow_slash : bool
-        Whether exactly one "/" character is allowed (this is reserved for 
-        parallel treebanks). This parameter replaces lcode, which was used to
-        allow slashes when equal to "ud".
-    known_sent_ids : set
-        The set of previously encountered sentence IDs.
-
-    Returns
-    -------
-    incidents : list
-        A list of Incidents (empty if validation is successful). 
-    """
-    incidents = []
-    matched = []
-    for c in comments:
-        match = crex.sentid.fullmatch(c)
-        if match:
-            matched.append(match)
-        else:
-            if c.startswith('# sent_id') or c.startswith('#sent_id'):
-                incidents.append(Error(
-                    testclass=TestClass.METADATA,
-                    level=2,
-                    testid='invalid-sent-id',
-                    message=f"Spurious sent_id line: '{c}' should look like '# sent_id = xxxxx' where xxxxx is not whitespace. Forward slash reserved for special purposes."
-                ))
-    if not matched:
-        incidents.append(Error(
-            testclass=TestClass.METADATA,
-            level=2,
-            testid='missing-sent-id',
-            message='Missing the sent_id attribute.'
-        ))
-    elif len(matched) > 1:
-        incidents.append(Error(
-            testclass=TestClass.METADATA,
-            level=2,
-            testid='multiple-sent-id',
-            message='Multiple sent_id attributes.'
-        ))
-    else:
-        # Uniqueness of sentence ids should be tested treebank-wide, not just file-wide.
-        # For that to happen, all three files should be tested at once.
-        sid = matched[0].group(1)
-        if sid in known_sent_ids:
-            incidents.append(Error(
-                testclass=TestClass.METADATA,
-                level=2,
-                testid='non-unique-sent-id',
-                message=f"Non-unique sent_id attribute '{sid}'."
-            ))
-        if sid.count('/') > 1 or (sid.count('/') == 1 and allow_slash):
-            incidents.append(Error(
-                testclass=TestClass.METADATA,
-                level=2,
-                testid='slash-in-sent-id',
-                message=f"The forward slash is reserved for special use in parallel treebanks: '{sid}'"
-            ))
-        #state.known_sent_ids.add(sid) # TODO: move this to the engine
-    return incidents
-
-def validate_text_meta(comments, tree, spaceafterno_in_effect):
-    """
-    Checks metadata other than sentence id, that is, document breaks, paragraph
-    breaks and sentence text (which is also compared to the sequence of the
-    forms of individual tokens, and the spaces vs. SpaceAfter=No in MISC).
-    """
-    incidents = []
-    newdoc_matched = []
-    newpar_matched = []
-    text_matched = []
-    for c in comments:
-        newdoc_match = crex.newdoc.fullmatch(c)
-        if newdoc_match:
-            newdoc_matched.append(newdoc_match)
-        newpar_match = crex.newpar.fullmatch(c)
-        if newpar_match:
-            newpar_matched.append(newpar_match)
-        text_match = crex.text.fullmatch(c)
-        if text_match:
-            text_matched.append(text_match)
-    if len(newdoc_matched) > 1:
-        incidents.append(Error(
-            testclass=TestClass.METADATA,
-            level=2,
-            testid='multiple-newdoc',
-            message='Multiple newdoc attributes.'
-        ))
-    if len(newpar_matched) > 1:
-        incidents.append(Error(
-			testclass=TestClass.METADATA,
-			level=2,
-            testid='multiple-newpar',
-            message='Multiple newpar attributes.'
-        ))
-    if (newdoc_matched or newpar_matched) and spaceafterno_in_effect:
-        incidents.append(Error(
-			testclass=TestClass.METADATA,
-			level=2,
-            testid='spaceafter-newdocpar',
-            message='New document or paragraph starts when the last token of the previous sentence says SpaceAfter=No.'
-        ))
-    if not text_matched:
-        incidents.append(Error(
-			testclass=TestClass.METADATA,
-			level=2,
-            testid='missing-text',
-            message='Missing the text attribute.'
-        ))
-    elif len(text_matched) > 1:
-        incidents.append(Error(
-			testclass=TestClass.METADATA,
-			level=2,
-
-            testid='multiple-text',
-            message='Multiple text attributes.'
-        ))
-    else:
-        stext = text_matched[0].group(1)
-        if stext[-1].isspace():
-            incidents.append(Error(
-			testclass=TestClass.METADATA,
-			level=2,
-                testid='text-trailing-whitespace',
-                message='The text attribute must not end with whitespace.'
-            ))
-        # Validate the text against the SpaceAfter attribute in MISC.
-        skip_words = set()
-        mismatch_reported = 0 # do not report multiple mismatches in the same sentence; they usually have the same cause
-        # We will sum state.sentence_line + iline, and state.sentence_line already points at
-        # the first token/node line after the sentence comments. Hence iline shall
-        # be 0 once we enter the cycle.
-        iline = -1
-        for cols in tree:
-            iline += 1
-            if 'NoSpaceAfter=Yes' in cols[utils.MISC]: # I leave this without the split("|") to catch all
-                incidents.append(Error(
-			testclass=TestClass.METADATA,
-			level=2,
-
-                    testid='nospaceafter-yes',
-                    message="'NoSpaceAfter=Yes' should be replaced with 'SpaceAfter=No'."
-                ))
-            if len([x for x in cols[utils.MISC].split('|') if re.match(r"^SpaceAfter=", x) and x != 'SpaceAfter=No']) > 0:
-                incidents.append(Error(
-			        testclass=TestClass.METADATA,
-			        level=2,
-                    # TODO: lineno=state.sentence_line+iline, (engine)
-                    testid='spaceafter-value',
-                    message="Unexpected value of the 'SpaceAfter' attribute in MISC. Did you mean 'SpacesAfter'?"
-                ))
-            if utils.is_empty_node(cols):
-                if 'SpaceAfter=No' in cols[utils.MISC]: # I leave this without the split("|") to catch all
-                    incidents.append(Error(
-			            testclass=TestClass.METADATA,
-			            level=2,
-                        # TODO: engine lineno=state.sentence_line+iline,
-                        testid='spaceafter-empty-node',
-                        message="'SpaceAfter=No' cannot occur with empty nodes."
-                    ))
-                continue
-            elif utils.is_multiword_token(cols):
-                beg, end = cols[utils.ID].split('-')
-                begi, endi = int(beg), int(end)
-                # If we see a multi-word token, add its words to an ignore-set – these will be skipped, and also checked for absence of SpaceAfter=No.
-                for i in range(begi, endi+1):
-                    skip_words.add(str(i))
-            elif cols[utils.ID] in skip_words:
-                if 'SpaceAfter=No' in cols[utils.MISC]:
-                    incidents.append(Error(
-			            testclass=TestClass.METADATA,
-			            level=2,
-                        # TODO: lineno=state.sentence_line+iline,
-                        testid='spaceafter-mwt-node',
-                        message="'SpaceAfter=No' cannot occur with words that are part of a multi-word token."
-                    ))
-                continue
-            else:
-                # Err, I guess we have nothing to do here. :)
-                pass
-            # So now we have either a multi-word token or a word which is also a token in its entirety.
-            if not stext.startswith(cols[utils.FORM]):
-                if not mismatch_reported:
-                    extra_message = ''
-                    if len(stext) >= 1 and stext[0].isspace():
-                        extra_message = ' (perhaps extra SpaceAfter=No at previous token?)'
-                    incidents.append(Error(
-			            testclass=TestClass.METADATA,
-			            level=2,
-                        # TODO: lineno=state.sentence_line+iline,
-                        testid='text-form-mismatch',
-                        message=f"Mismatch between the text attribute and the FORM field. Form[{cols[utils.ID]}] is '{cols[utils.FORM]}' but text is '{stext[:len(cols[utils.FORM])+20]}...'"+extra_message
-                    ))
-                    mismatch_reported = 1
-            else:
-                stext = stext[len(cols[utils.FORM]):] # eat the form
-                # Remember if SpaceAfter=No applies to the last word of the sentence.
-                # This is not prohibited in general but it is prohibited at the end of a paragraph or document.
-                if 'SpaceAfter=No' in cols[utils.MISC].split("|"):
-                    spaceafterno_in_effect = True
-                else:
-                    spaceafterno_in_effect = False
-                    if (stext) and not stext[0].isspace():
-                        incidents.append(Error(
-			                testclass=TestClass.METADATA,
-			                level=2,
-                            # TODO: lineno=state.sentence_line+iline,
-                            testid='missing-spaceafter',
-                            message=f"'SpaceAfter=No' is missing in the MISC field of node {cols[utils.ID]} because the text is '{utils.shorten(cols[utils.FORM]+stext)}'."
-                        ))
-                    stext = stext.lstrip()
-        if stext:
-            incidents.append(Error(
-			    testclass=TestClass.METADATA,
-			    level=2,
-                testid='text-extra-chars',
-                message=f"Extra characters at the end of the text attribute, not accounted for in the FORM fields: '{stext}'"
-            ))
